pyproject.toml
[project]
name = "corp-profile"
version = "0.1.0"
description = "Build rich company context documents from corp-graph Postgres"
requires-python = ">=3.13"
dependencies = [
    "psycopg[binary]>=3.2.0",
    "pydantic>=2.12.0",
    "python-dotenv>=1.2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
